requirements.txt
requests
futures
google-cloud-storage
google-cloud-pubsub